core/requirements.txt
ortools
pandas
ruamel.yaml
munch
pytest
